# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Hospital Management Solution",
      "version": "1.0.0",
      "title": "Hospital Management Solution API documentation",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "servers": [
      {
        "url": "/api/"
      }
    ],
...
    "paths": {
      "/patient/add-patient-info/": {
        "post": {
          "tags": [
            "Patients Info"
          ],
          "summary": "Create a new patient info record",
          "requestBody": {
            "description": "Create a new patient info record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/patientAddInfo"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": ""
            }
          }
        }
      },
      "/patient/fetch-patient-info/": {
        "get": {
          "tags": [
            "Patients Info"
          ],
          "summary": "Fetch patients info record of all patients",
          "parameters": [
            {
                "name": "page_no",
                "in": "query",
                "description": "Page number",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 0
              },
              {
                "name": "page_size",
                "in": "query",
                "description": "Page size",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 1
              },
              {
                "name": "patient_email",
                "in": "query",
                "description": "filter by email value",
                "required": true,
                "type": "string",
                "example": "user@example"
              }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/patient/fetch-patient-info/{patient_id}": {
        "get": {
          "tags": [
            "Patients Info"
          ],
          "summary": "Fetch patients info record of a particular patient by id",
          "parameters": [
            {
                "name": "patient_id",
                "in": "path",
                "description": "ID of the patient to fetch",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 2
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/patient/modify-patient-info/{patient_id}": {
        "put": {
          "tags": [
            "Patients Info"
          ],
          "summary": "Modify patient information by ID",
          "parameters": [
            {
              "name": "patient_id",
              "in": "path",
              "description": "ID of the patient to modify",
              "required": true,
              "type": "integer",
              "format": "int32",
              "example": 2
            }
          ],
          "requestBody": {
            "description": "Edit a patient info record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/patientEditInfo"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "",
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "/patient/add-patient-history/": {
        "post": {
          "tags": [
            "Patients History"
          ],
          "summary": "Create a new patient history record",
          "requestBody": {
            "description": "Create a new patient history record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/patientAddHistoryInfo"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": ""
            }
          }
        }
      },
      "/patient/fetch-patient-history/{patient_id}": {
        "get": {
          "tags": [
            "Patients History"
          ],
          "summary": "Fetch patients history record of a particular patient by id",
          "parameters": [
            {
                "name": "patient_id",
                "in": "path",
                "description": "ID of the patient to fetch",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 2
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/patient/modify-patient-history/{patient_id}": {
        "put": {
          "tags": [
            "Patients History"
          ],
          "summary": "Modify patient history information by ID",
          "parameters": [
            {
              "name": "patient_id",
              "in": "path",
              "description": "ID of the patient to modify",
              "required": true,
              "type": "integer",
              "format": "int32",
              "example": 2
            }
          ],
          "requestBody": {
            "description": "Edit a patient history record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/patientEditHistoryInfo"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "",
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "/patient/add-patient-appointment/": {
        "post": {
          "tags": [
            "Patients Appointment"
          ],
          "summary": "Create a new patient history appointment record",
          "requestBody": {
            "description": "Create a new patient appointment record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/patientAddAppointment"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": ""
            }
          }
        }
      },
      "/patient/fetch-patient-appointment/{patient_id}": {
        "get": {
          "tags": [
            "Patients Appointment"
          ],
          "summary": "Fetch patients appointment record of a particular patient by id",
          "parameters": [
            {
                "name": "patient_id",
                "in": "path",
                "description": "ID of the patient to fetch",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 1
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/patient/modify-patient-appointment/{patient_id}": {
        "put": {
          "tags": [
            "Patients Appointment"
          ],
          "summary": "Modify patient appointment information by ID",
          "parameters": [
            {
              "name": "patient_id",
              "in": "path",
              "description": "ID of the patient to modify",
              "required": true,
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          ],
          "requestBody": {
            "description": "Edit a patient appointment record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/patientEditAppointment"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "",
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "patientAddInfo": {
          "type": "object",
          "properties": {
            "gender": {
              "type": "string",
              "format": "string",
              "example": "Male"

            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "age": {
                "type": "number",
                "format": "number",
                "example": 40
            },
            "name": {
                "type": "string",
                "format": "string",
                "example": "John Doe"
            },
            "phone": {
                "type": "string",
                "format": "string",
                "example": "+91 12345678"
            }
          }
        },
...
        "patientEditInfo": {
            "type": "object",
          "properties": {
            "gender": {
              "type": "string",
              "format": "string",
              "example": "Female"

            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "age": {
                "type": "number",
                "format": "number",
                "example": 30
            },
            "name": {
                "type": "string",
                "format": "string",
                "example": "Doe John"
            },
            "phone": {
                "type": "string",
                "format": "string",
                "example": "+91 8765432"
            }
          }
        },
        "patientAddHistoryInfo": {
            "type": "object",
            "properties": {
              "diagnosis": {
                "type": "string",
                "format": "string",
                "example": "The patient is having mild fever."
  
              },
              "doctor_id": {
                  "type": "number",
                  "format": "number",
                  "example": 1
              },
              "patient_id": {
                "type": "number",
                "format": "number",
                "example": 1
              },
              "appointment_time": {
                "type": "number",
                "format": "number",
                "example": 1711813215
              },
              "allergies": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["Peanut", "Jaggery"]
              },
              "medications": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["Isobufren", "Isomorphin"]
              }
            }
        },
        "patientEditHistoryInfo": {
            "type": "object",
            "properties": {
              "diagnosis": {
                "type": "string",
                "format": "string",
                "example": "The patient is having high fever."
  
              },
              "doctor_id": {
                  "type": "number",
                  "format": "number",
                  "example": 1
              },
              "appointment_time": {
                "type": "number",
                "format": "number",
                "example": 1711813192
              },
              "allergies": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["Sugar"]
              },
              "medications": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["Azithromycin"]
              }
            }
        },
        "patientAddAppointment": {
            "type": "object",
            "properties": {
              "doctor_id": {
                  "type": "number",
                  "format": "number",
                  "example": 1
              },
              "patient_id": {
                "type": "number",
                "format": "number",
                "example": 1
              },
              "appointment_time": {
                "type": "number",
                "format": "number",
                "example": 1711813215
              }
            }
        },
        "patientEditAppointment": {
            "type": "object",
            "properties": {
              "doctor_id": {
                  "type": "number",
                  "format": "number",
                  "example": 1
              },
              "patient_id": {
                "type": "number",
                "format": "number",
                "example": 1
            },
              "appointment_time": {
                "type": "number",
                "format": "number",
                "example": 1711813192
              }
            }
        },
...
      }
    }
  }